Makes sure recycled HANDLEs are replaced in cache
In GlobalContext::UnlockRect() we were checking for the existence of the
replaced HANDLE in the cache and not inserting it if it was already there.
The problem is that since the game recycles HANDLEs we need to go ahead
and insert the entry in case it's the same HANDLE but a different hash and
thus a different replacement texture. TextureCache::map_insert() already
makes sure there are not duplicate entries by checking the new hash
against the existing hash before adding a cache entry.

Also when a file fails to load we need to be sure to call
TextureCache::erase(HANDLE) to clear any existing entries.

This should solve the issue where in-game texture were being replaced with
the wrong replacement image.

Also adds DEBUG preprocessor macro for toggling debug logging to
texture_cache.log; to turn on debugging simply change DEBUG from 0 to 1.