Fixed issue with unset debugtype
If debugtype is unset then it resolves to "", which causes debug images to
be saved to \tonberry\debug\. This is extremely undesirable behavior, but
was occurring whenever a texture was replaced.

We also had an issue where hash collisions that caused Hash_Algorithm_2 to
run and then ended up as "NO_MATCH2" still did a hash lookup in the
TextureCache at the if(!texcache->update(Handle, hash)) line. To avoid
this, I set a "match" boolean that is initiated as true and only becomes
false in the event of a collision and ensuing NO_MATCH2. If match is true,
the debugtype is set to "replaced" and only then do we call
texcache->update().